Fix key comparison
This oversight caused corruption of the recent-headings list,
overwriting old headings with newer ones.  No data is lost, of course,
because it's just a cache, but the list will have to essentially start
from scratch.

Hopefully this will be the last sneaky bug, allowing a stable version to
be tagged soon.